+++ src/main/java/com/vonage/client/messages/viber/ViberTextRequest.java
@@ -22,7 +22,7 @@ import com.vonage.client.messages.internal.Text;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 public final class ViberTextRequest extends ViberRequest {
-	String text;
+	final String text;
 
 	ViberTextRequest(Builder builder) {
 		super(builder, MessageType.TEXT);